# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAR Fleet Service",
    "summary": "Ground fleet service for civilian search-and-rescue drone operations.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Exchange a username and password for a session token.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts; see Retry-After.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "End the current session.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-permission": "authenticated"
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "Who am I, and what may I do.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-permission": "authenticated"
      }
    },
    "/api/v1/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "description": "Change one's own password; every other session of this user is revoked.",
        "operationId": "change_password_api_v1_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-permission": "authenticated"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc",
            "description": "Where the value is: body, query, path, header."
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "description": "One invalid input value."
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Credentials. The username is case-insensitive."
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "token",
          "token_type",
          "expires_at",
          "user",
          "permissions"
        ],
        "title": "LoginResponse",
        "description": "A new session. The token is shown only here; send it as ``Authorization: Bearer``."
      },
      "MeResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "type": "array",
            "title": "Permissions"
          },
          "session_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Session Expires At"
          }
        },
        "type": "object",
        "required": [
          "user",
          "permissions",
          "session_expires_at"
        ],
        "title": "MeResponse",
        "description": "The current session's user and what they may do."
      },
...
      "PasswordChange": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 10,
            "title": "New Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange",
        "description": "Change one's own password; other sessions of the user are revoked."
      },
...
      "Permission": {
        "type": "string",
        "enum": [
          "fleet.view",
          "missions.plan",
          "fleet.manage",
          "incidents.manage",
          "geofences.manage",
          "users.view",
          "audit.read",
          "users.manage"
        ],
        "title": "Permission",
        "description": "Something a session may be allowed to do."
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank",
            "examples": [
              "urn:sar-gcs:problem:conflict"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Present on validation errors: each invalid value."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details. Extension members depend on ``type``."
      },
      "Role": {
        "type": "string",
        "enum": [
          "observer",
          "operator",
          "supervisor",
          "admin"
        ],
        "title": "Role",
        "description": "Operator roles, lowest to highest authority (ADR 0009)."
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "is_active",
          "created_at",
          "updated_at",
          "last_login_at"
        ],
        "title": "UserOut",
        "description": "An operator account (never includes credentials)."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Session token returned by POST /api/v1/auth/login.",
        "scheme": "bearer"
      }
    }
  },
...
}